requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
cloudinary==1.36.0
colorama==0.4.6
dj-database-url==2.1.0
Django==4.2.5
django-ipware==5.0.2
frozenlist==1.4.0
geoip2==4.7.0
gunicorn==21.2.0
idna==3.4
maxminddb==2.4.0
multidict==6.0.4
numpy==1.26.1
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
whitenoise==6.6.0
yarl==1.9.2
django-environ==0.10.0
google-cloud-secret-manager==2.16.1
